+++ public/data/list.json
@@ -1,31 +1,69 @@
 [
   {
+    "id": 1,
     "title": "콜드스퀘어",
-    "category_type": "원룸",
-    "trade_type": "월세",
     "latitude": 37.62197524055062,
     "longitude": 127.16017523675508,
+    "category_type": "원룸",
     "supply_size": 34.4,
     "current_floor": "3층",
     "price_main": 0,
     "price_deposit": 500,
     "price_monthly": 35,
-    "description_title": "콜드스퀘어 3층 깔끔 개별난방 싸진않음 햇빛은 잘들어옴 입주협의"
+    "description_title": "콜드스퀘어 3층 깔끔 개별난방 싸진않음 햇빛은 잘들어옴 입주협의",
+    "image_url": "https://d1774jszgerdmk.cloudfront.net/512/954acfb5-256d-4d74-a4db-8f43ca08a3d5",
+    "ableDate": "즉시입주",
+    "tradeTypes": ["월세"],
+    "isLike": 1
   },
   {
+    "id": 2,
     "title": "하남돼지집",
     "latitude": 37.620842424005616,
-    "longitude": 127.1583774403176
+    "longitude": 127.1583774403176,
+    "category_type": "빌라",
+    "supply_size": 64.4,
+    "current_floor": "5층",
+    "price_main": 10000,
+    "price_deposit": 500,
+    "price_monthly": 60,
+    "description_title": "하남돼지집 윗층에 사람사는곳 있습니다",
+    "image_url": "https://d1774jszgerdmk.cloudfront.net/512/08f50010-a583-4fac-9793-4743546d0af5",
+    "ableDate": "즉시입주",
+    "tradeTypes": ["전세", "월세"],
+    "isLike": 1
   },
   {
+    "id": 3,
     "title": "수유리우동",
     "latitude": 37.624915253753194,
-    "longitude": 127.15122688059974
+    "longitude": 127.15122688059974,
+    "category_type": "아파트",
+    "supply_size": 94.4,
+    "current_floor": "3층",
+    "price_main": 38000,
+    "price_deposit": 500,
+    "price_monthly": 35,
+    "description_title": "상갓집인줄 알았겠지만 수유리우동은 아파트 이름입니다",
+    "image_url": "https://d1774jszgerdmk.cloudfront.net/512/ee54045f-f361-4990-bb02-82d2948aa74f",
+    "ableDate": "즉시입주",
+    "tradeTypes": ["전세"]
   },
   {
+    "id": 4,
     "title": "맛닭꼬",
     "latitude": 37.62456273069659,
-    "longitude": 127.15211256646381
+    "longitude": 127.15211256646381,
+    "category_type": "아파트",
+    "supply_size": 94.4,
+    "current_floor": "10층",
+    "price_main": 20000,
+    "price_deposit": 500,
+    "price_monthly": 35,
+    "description_title": "전세 0억 로직이 너무 지저분한데; 제 코드 줄여주실분?",
+    "image_url": "https://d1774jszgerdmk.cloudfront.net/512/681f27c7-3a4f-4a54-b665-ff2fee0addd0",
+    "ableDate": "즉시입주",
+    "tradeTypes": ["월세"]
   },
   { "latitude": 37.27943075229118, "longitude": 127.01763998406159 },
   { "latitude": 37.55915668706214, "longitude": 126.92536526611102 },
