+++ src/pages/map-page/List.js
@@ -1,114 +1,25 @@
-import React, { useContext, useState } from 'react';
+import React, { useContext } from 'react';
 import styled from 'styled-components';
-import { RealEstateContext, RealEstateContextDispatch } from './context';
-import { IoMdHeartEmpty, IoMdHeart } from 'react-icons/io';
+import { RealEstateContext } from './context';
+import ListCard from './ListCard';
 
 function List() {
   const RealEstate = useContext(RealEstateContext);
-  const RealEstateDispatch = useContext(RealEstateContextDispatch);
-  const [like, setLike] = useState(false);
+  // console.log(RealEstate.realEstate);
 
-  const LikeHandler = () => {
-    setLike(!like);
-  };
-  // console.log(RealEstate);
-  // console.log(RealEstateDispatch);
-  // console.log(RealEstate.selected[0]);
-  // console.log(RealEstate.selected[0].description_title);
   return (
-    <>
-      {RealEstate.selected.length === 0 ? null : (
-        <ListWrapper>
-          <CardWrapper>
-            <Card>
-              <ImageWrapper>
-                <Like onClick={LikeHandler}>
-                  {like === false ? (
-                    <IoMdHeartEmpty color="black" />
-                  ) : (
-                    <IoMdHeart color="red" />
-                  )}
-                </Like>
-              </ImageWrapper>
-              <InfoWrap>
-                <p class="price">
-                  {RealEstate.selected[0].trade_type === '월세'
-                    ? `${RealEstate.selected[0].trade_type} 
-                      ${RealEstate.selected[0].price_deposit}/${RealEstate.selected[0].price_monthly}`
-                    : RealEstate.selected[0].trade_type}
-                </p>{' '}
-                <br />
-                <p class="type">{RealEstate.selected[0].trade_type}</p>
-                <br />
-                <p class="description">
-                  {RealEstate.selected[0].current_floor}, &nbsp;
-                  {RealEstate.selected[0].supply_size}
-                  <br />
-                  {RealEstate.selected[0].description_title}
-                </p>
-              </InfoWrap>
-            </Card>
-          </CardWrapper>
-        </ListWrapper>
-      )}
-    </>
+    <Wrapper>
+      {RealEstate.selected.length === 0
+        ? null
+        : RealEstate.selected.map(data => (
+            <ListCard key={data.id} data={data} />
+          ))}
+    </Wrapper>
   );
 }
 
 export default List;
 
-const ListWrapper = styled.div`
-  // background-color: yellow;
+const Wrapper = styled.div`
   margin-top: 65px;
-  border-bottom: 1px solid rgb(231, 231, 231);
-  transition: 0.3s;
-  :hover {
-    background-color: #f0f4fa;
-  }
-`;
-const CardWrapper = styled.div`
-  border: 1px solid transparent;
-`;
-const Card = styled.div`
-  // border: 1px solid blue;
-  display: flex;
-  justify-content: flex-start;
-  margin: 1rem;
-`;
-const ImageWrapper = styled.div`
-  border: 1px solid blue;
-  position: relative;
-  width: 140px;
-  height: 140px;
-`;
-const Like = styled.div`
-  position: absolute;
-  top: 5px;
-  right: 5px;
-  color: ${props => props.color};
-  font-size: 1.5rem;
-  cursor: pointer;
-`;
-const InfoWrap = styled.div`
-  // border: 1px solid black;
-  margin-left: 1rem;
-  padding-top: 0.2rem;
-  width: 13rem;
-
-  .price {
-    font-size: 1.2rem;
-    font-weight: bold;
-  }
-  .type {
-    font-size: 0.8rem;
-    font-weight: 500;
-  }
-  .description {
-    font-size: 0.9rem;
-    font-weight: 350;
-    line-height: 1.1rem;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-  }
 `;
